+++ app/controllers/all/projects_controller.rb
@@ -10,6 +10,14 @@ class All::ProjectsController < ApplicationController
     pre_fetch_incoming_trusts(@projects)
   end
 
+  def with_academy_urn
+    authorize Project, :index?
+    @pager, @projects = pagy(Project.with_academy_urn.by_conversion_date)
+
+    pre_fetch_establishments(@projects)
+    pre_fetch_incoming_trusts(@projects)
+  end
+
   private def pre_fetch_establishments(projects)
     EstablishmentsFetcher.new.call(projects)
   end